feat(batch-e): convert support-chat, chat, sms, scheduled-matching, lab, moment, force-matching to v2
