+++ app/admin/moment/page.tsx
@@ -1,49 +1,14 @@
-'use client';
+import { getRouteMode, isAdminShellV2Enabled } from '@/shared/feature-flags';
+import MomentManagementLegacy from './moment-legacy';
+import MomentManagementV2 from './moment-v2';
 
-import { useState } from 'react';
-import {
-  Box,
-  Typography,
-  Tabs,
-  Tab,
-} from '@mui/material';
-import QuestionGenerationTab from './components/QuestionGenerationTab';
-import QuestionListTab from './components/QuestionListTab';
-import QuestionTranslationTab from './components/QuestionTranslationTab';
-import { LegacyPageAdapter } from '@/shared/ui/admin/legacy-page-adapter';
+export default async function MomentManagementPage() {
+  const shellV2 = await isAdminShellV2Enabled();
+  const mode = await getRouteMode('moment');
 
-type TabValue = 'generation' | 'list' | 'translation';
+  if (shellV2 && mode === 'v2') {
+    return <MomentManagementV2 />;
+  }
 
-function MomentManagementPageContent() {
-  const [tabValue, setTabValue] = useState<TabValue>('list');
-
-  const handleTabChange = (_: React.SyntheticEvent, newValue: TabValue) => {
-    setTabValue(newValue);
-  };
-
-  return (
-    <Box>
-      <Typography variant="h5" fontWeight="bold" sx={{ mb: 3 }}>
-        모먼트 관리
-      </Typography>
-
-      <Tabs value={tabValue} onChange={handleTabChange} sx={{ mb: 3 }}>
-        <Tab label="질문 목록" value="list" />
-        <Tab label="질문 생성" value="generation" />
-        <Tab label="질문 번역" value="translation" />
-      </Tabs>
-
-      {tabValue === 'list' && <QuestionListTab />}
-      {tabValue === 'generation' && <QuestionGenerationTab />}
-      {tabValue === 'translation' && <QuestionTranslationTab />}
-    </Box>
-  );
-}
-
-export default function MomentManagementPage() {
-  return (
-    <LegacyPageAdapter>
-      <MomentManagementPageContent />
-    </LegacyPageAdapter>
-  );
+  return <MomentManagementLegacy />;
 }
